requirements.txt
FastAPI
paho-mqtt >= 2
python-dotenv
# uvicorn